# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: SMS Blossom API
  version: 1.0.0
  description: >
    REST API for Shopify SMS marketing app (Express, ESM).
    Includes OAuth-driven backend, consent flows, discounts, reporting, webhooks, and GDPR.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.sms-blossom.com
    description: Production API
...
security: []
paths:
...
  /discounts:
    post:
      tags: [Discounts]
      summary: Create a discount code (Shopify Admin GraphQL)
      operationId: createDiscount
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DiscountCreateRequest' }
      responses:
        '200':
          description: Discount created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DiscountCreateResponse' }
        '400':
          description: Missing shop parameter
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Code conflict
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Invalid payload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /discounts/apply-url:
    get:
      tags: [Discounts]
      summary: Build canonical apply URL for a code
      operationId: buildApplyUrl
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
        - in: query
          name: code
          required: true
          schema: { type: string }
        - in: query
          name: redirect
          required: false
          schema: { type: string, default: /cart }
      responses:
        '200':
          description: URL built
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApplyUrlResponse' }
        '400':
          description: Missing required parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /campaigns/{id}/snapshot:
    post:
      tags: [Campaigns]
      summary: Snapshot audience from segment/filter into recipients
      operationId: snapshotCampaign
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: shop
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Snapshot done
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  total: { type: integer }
                  inserted: { type: integer }
        '404':
          description: Unknown shop or campaign
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /campaigns/{id}/estimate:
    get:
      tags: [Campaigns]
      summary: Estimate cost/segments for pending recipients
      operationId: estimateCampaign
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: shop
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Estimate returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  recipients: { type: integer }
                  perMessageSegments: { type: integer }
                  totalSegments: { type: integer }
                  estCost: { type: number }
        '404':
          description: Unknown shop or campaign
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /campaigns/{id}/test-send:
    post:
      tags: [Campaigns]
      summary: Test send to a specific phone (E.164)
      operationId: testSendCampaign
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: shop
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone: { type: string }
              required: [phone]
      responses:
        '200':
          description: Test send result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  result: { type: object }
        '404':
          description: Unknown shop, campaign, or contact
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Missing phone
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /campaigns/{id}/send-now:
    post:
      tags: [Campaigns]
      summary: Send to snapshotted audience (batched, throttled)
      operationId: sendCampaign
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: shop
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Send summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  sent: { type: integer }
                  failed: { type: integer }
                  skipped: { type: integer }
        '404':
          description: Unknown shop or campaign
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /campaigns/{id}/attach-discount:
    post:
      tags: [Campaigns]
      summary: Attach an existing Discount to a Campaign
      operationId: attachDiscountToCampaign
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: shop
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                discountId: { type: string }
                code: { type: string }
      responses:
        '200':
          description: Attached
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  campaignId: { type: string }
                  discountId: { type: string }
        '404':
          description: Unknown shop, campaign, or discount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /campaigns/{id}/detach-discount:
    post:
      tags: [Campaigns]
      summary: Detach discount from Campaign
      operationId: detachDiscountFromCampaign
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: shop
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Detached
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        '404':
          description: Unknown shop or campaign
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /campaigns/{id}/utm:
    put:
      tags: [Campaigns]
      summary: Set UTM params for Campaign
      operationId: setCampaignUtm
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: shop
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  utm: { type: object }
        '404':
          description: Unknown shop or campaign
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /campaigns/{id}/apply-url:
    get:
      tags: [Campaigns]
      summary: Preview campaign discount apply URL (UTM/shortlink)
      operationId: getCampaignApplyUrl
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: query
          name: shop
          required: true
          schema: { type: string }
        - in: query
          name: redirect
          schema: { type: string, default: '/checkout' }
        - in: query
          name: short
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: URL returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  url: { type: string }
                  short: { type: [string, 'null'] }
        '404':
          description: Unknown shop or campaign
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '422':
          description: Campaign has no discount
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /discounts/conflicts:
    get:
      tags: [Discounts]
      summary: Advisory scan for active automatic discounts (possible conflicts)
      operationId: getDiscountConflicts
      parameters:
        - in: query
          name: shop
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Active automatic discounts returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  automaticDiscounts: { type: array, items: { type: object } }
        '400':
          description: Missing shop or unknown shop
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  schemas:
...
    DiscountCreateRequest:
      type: object
      properties:
        code: { type: string, example: 'WELCOME10' }
        title: { type: string }
        kind: { type: string, enum: [percentage, amount] }
        value: { type: number, example: 10 }
        currencyCode: { type: string, example: 'EUR' }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        appliesOncePerCustomer: { type: boolean, default: true }
        usageLimit: { type: integer }
        redirect: { type: string, example: '/checkout' }
        segments:
          type: array
          items: { type: string }
      required: [code, kind, value]
    DiscountCreateResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        code: { type: string, example: 'WELCOME10' }
        title: { type: string }
        id: { type: string }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        applyUrl:
          {
            type: string,
            example: 'https://shop.myshopify.com/discount/WELCOME10?redirect=%2Fcheckout',
          }
    ApplyUrlResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        url: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        details: {}